+++ src/javascript/components/explorer/item.js
@@ -5,6 +5,8 @@ import Tab from '../../constructors/tab'
 import Editor from '../../constructors/editor'
 import path from 'path'
 import newDirectoryDialog from '../../defaults/dialogs/new.directory'
+import StaticConfig from 'StaticConfig'
+import ThemeRegistry from 'ThemeRegistry'
 
 import ThemeProvider from 'ThemeProvider'
 import Icons from '../../../../assets/icons/**.svg'
@@ -88,7 +90,7 @@ function Item(){
                         new Editor({
                             language:fileExtension,
                             value:data ,
-                            theme:'Arctic',
+                            theme:ThemeRegistry.registry.data.list[StaticConfig.data.theme].textTheme,
                             bodyElement,
                             tabElement,
                             panel