+++ src/javascript/utils/extension.loader.js
@@ -1,9 +1,15 @@
-import Window from '../constructors/window'
-import Menu from '../constructors/menu'
-import {puffin} from '@mkenzo_8/puffin'
+import { puffin } from '@mkenzo_8/puffin'
 
-function loadPlugin(path){
-    require(path).main({Window,puffin,Menu})
+function loadExtension(path){
+   return require(path).main({Window,puffin,Menu})
 }
 
-export default loadPlugin
+function initExtension(extension){
+    return extension.entry()
+}
+
+export { loadExtension, initExtension }
+
+/**
+ * entry() ->  Initial extension's function, called when the plugin is executed
+ */